requirements.txt
pandas>=1.5.0
streamlit>=1.25.0
matplotlib>=3.5.0
seaborn>=0.11.0
openai>=1.0.0
jinja2>=3.0.0
pytest>=7.0.0
python-dotenv>=1.0.0
requests-mock>=1.11.0 